fix: run init after modules load to enable join
