+++ static/index.html
@@ -142,6 +142,6 @@
   <!-- Тосты -->
   <div id="toasts" class="toasts" aria-live="polite"></div>
 
-  <script src="/app.js"></script>
+  <script type="module" src="/app.js"></script>
 </body>
 </html>
